Added/Fixed Sounds and Typos
- Fixed all typos on trivia?
- Added new incorrect sound
- Added main menu buttons graphics

ALSO I FOUND A BUG WHERE YOU DO NOT DIE AT  0 LIVES. I played from the main menu and had just answered a question correctly. I tried to self-destruct on a cluster of asteroids, but my life count went to the negatives. I couldn't replicate it, so watch out for that.